Telegram Desktop/saas-ag9/server/requirements.txt
fastapi
uvicorn
google-auth
requests
python-multipart
pydantic
PyJWT
prometheus_client
motor
stripe